+++ inlong-sort/sort-common/src/main/java/org/apache/inlong/sort/protocol/node/ExtractNode.java
@@ -73,6 +73,10 @@ public abstract class ExtractNode implements Node {
 
     public static final String INLONG_MSG = "inlong-msg";
 
+    public static final String INLONG_MSG_AUDIT_TIME = "value.data-time";
+
+    public static final String CONSUME_AUDIT_TIME = "consume_time";
+
     @JsonProperty("id")
     private String id;
     @JsonInclude(Include.NON_NULL)
